# Simplify and speed up analyser highway_vs_buildings