@Famlam 2023-05-20
FYI, the jumping around of markers doesn't seem to be related to this PR and happens between Osmose runs (live on the server). I downloaded (on 18-05 and 20-05) [the json file with issues of usa_california_kern of class 4,5,8,9](https://osmose.openstreetmap.fr/api/0.3/issues?item=1070&class=4,5,8,9&limit=10000&country=usa_california_kern) and compared the output. It turns out that quite a few markers are "jumping around". Some examples:

`>` = run of Osmose on May 18
`<` = run of Osmose on May 20

```
> { "lat": 34.822653, "lon": -117.6807164, "id": "765034bb-9228-180f-49c1-6b40eaffe47d", "item": "1070"},
< { "lat": 34.822661, "lon": -117.6815496, "id": "765034bb-9228-180f-49c1-6b40eaffe47d", "item": "1070"},

> { "lat": 35.087376, "lon": -118.0007731, "id": "b69ca5fa-ce99-b850-3234-3e817764bbfd", "item": "1070"},
< { "lat": 35.087417, "lon": -117.9970399, "id": "b69ca5fa-ce99-b850-3234-3e817764bbfd", "item": "1070"},

> { "lat": 35.097527, "lon": -117.6652877, "id": "c6c2bac9-9784-6917-6f62-a04ecb4838f5", "item": "1070"},
< { "lat": 35.097556, "lon": -117.6632715, "id": "c6c2bac9-9784-6917-6f62-a04ecb4838f5", "item": "1070"},

> { "lat": 35.1659434, "lon": -119.2569257, "id": "38310e6d-b96e-9f9b-0cb2-782fbe796c49", "item": "1070"},
< { "lat": 35.1660004, "lon": -119.2656135, "id": "38310e6d-b96e-9f9b-0cb2-782fbe796c49", "item": "1070"},

(and several others)
```

For the last one I verified that no nodes or ways have been modified in the past month.