@Famlam 2023-05-27
The majority of the results look ok, but it seems that dropping the 100 nodes restriction has exploded the item=1070 class=4 (highway intersecting minor stream) results in canada_british_columbia and a few usa states. 
https://osmose.openstreetmap.fr/nl/issues/open?item=1070&class=4&country=canada_british_columbia
Not sure if there's much we can do here though. The majority of the streams I cannot even see on satellite imagery, so I guess these are either really small or rainfall-only streams. 
Maybe we should group them together per stream (e.g. array_agg on highway.id) so they appear as single markers per stream?