@frodrigo 2023-05-27
We already disabled osmosis_waterway due to too many issues in Canada. Maybe we can also disable this one, at least only for streams ?

https://github.com/osm-fr/osmose-backend/blob/dev/osmose_config.py#L1088

https://osmose.openstreetmap.fr/en/map/#item=1070&class=4&zoom=10&lat=48.3014&lon=-77.2799&level=1%2C2%2C3

cc @matthewdarwin