@Famlam 2023-05-28
> Maybe we can also disable this one, at least only for streams ?

Hmm, should be doable. Only CA-BC I assume, the other Canada regions weren't so much affected. Would you like to do this now or only when a user report comes in?

In addition, since the many-reports-per-`waterway=stream` seems to be a global thing, where typically a single stream is not detailed with `bridge=*` or `ford=*` throughout the length of it ([example Washington](https://osmose.openstreetmap.fr/nl/map/#source=10646&item=1070&class=4&zoom=13&lat=46.4045&lon=-120.4638&level=1%2C2%2C3), [example Oregon](https://osmose.openstreetmap.fr/nl/map/#source=12015&item=1070&class=4&zoom=14&lat=44.05925&lon=-122.33166&level=1%2C2%2C3), [example Portugal](https://osmose.openstreetmap.fr/nl/map/#source=6863&item=1070&class=4&zoom=16&lat=40.84445&lon=-8.43936&level=1%2C2%2C3), [example Italy](https://osmose.openstreetmap.fr/nl/map/#source=2928&item=1070&class=4&zoom=16&lat=43.35175&lon=10.86401&level=1%2C2%2C3), [example Taiwan](https://osmose.openstreetmap.fr/nl/map/#source=22227&item=1070&class=4&zoom=18&lat=24.469875&lon=120.967631&level=1%2C2%2C3), [example Brazil](https://osmose.openstreetmap.fr/nl/map/#source=36876&item=1070&class=4&status=false&zoom=16&lat=-7.96266&lon=-45.08702&level=1%2C2%2C3)) do you think it's also a good idea to group them together into a single marker? Or preferably not?